+++ module-checkout-ui/_run.php
@@ -9,7 +9,9 @@ add_action( 'template_redirect', '_run_h_checkout_ui' );
 /////
 
 function _run_h_checkout_ui() {
-  if( is_checkout() && get_theme_support('h-wc-checkout') ) {
+  if( !get_theme_support('h-checkout') || !get_theme_support('h-wc-checkout') ) { return false; }
+
+  if( is_checkout() ) {
     require_once HOO_PATH . '/module-checkout-ui/checkout-ui.php';
     require_once HOO_PATH . '/module-checkout-ui/form-fields.php';
     new \h\Checkout_UI();